Move the existing modules into there own folder
If we are going to start splitting stuff, it makes sense to add them to their own folder rather than adding yet more to the main folder.
